Fix bugs and cleanups suggested by clang-tidy

+++ src/gui_editbox.cpp
@@ -985,7 +985,8 @@ void edit_box::update_carret_position_()
                 (iterCarretPos_ - sUnicodeText_.begin()) - uiDisplayPos_;
         }
 
-        float fYOffset = (pText->get_num_lines() - 1) * (pText->get_line_height() * pText->get_line_spacing());
+        float fYOffset = static_cast<float>(
+            (pText->get_num_lines() - 1) * (pText->get_line_height() * pText->get_line_spacing()));
 
         std::size_t uiIndex = iterDisplayCarret - sDisplayedText_.begin();
 
@@ -1021,6 +1022,7 @@ bool edit_box::add_char_(char32_t sUnicode)
     if (get_num_letters() >= uiMaxLetters_)
         return false;
 
+    // TODO: use localizer for these checks, if possible
     if (bNumericOnly_)
     {
         if (sUnicode == U'.')
